# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Exam Project",
    "description": "Documentation Exam Project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/admin/login": {
      "post": {
        "tags": [
          "Auth-Admin"
        ],
        "description": "Login in to account. It will return a token, admin's id and admin's username",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains email and password, required to log in",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth-User"
        ],
        "description": "Login in to account. It will return a token, user's id and user's username",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains email and password, required to log in",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth-User"
        ],
        "description": "Sign up user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains credentials such as fristName, lastName, username, email, password, address and phoneNumber, required to sign up",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignupUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/auth/login": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "adminToken": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "LoginUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "Mohsen@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "0000"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "SignupUser": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Mohsen"
        },
        "lastName": {
          "type": "string",
          "example": "Mohammadi"
        },
        "username": {
          "type": "string",
          "example": "2050"
        },
        "email": {
          "type": "string",
          "example": "Mohsen@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "0000"
        },
        "address": {
          "type": "string",
          "example": "stavanger 4017"
        },
        "phoneNumber": {
          "type": "string",
          "example": "344556"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "username",
        "email",
        "password",
        "address",
        "phoneNumber"
      ]
    },
    "LoginAdmin": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "admin@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "0000"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
...
  }
}